api_notas_fiscais/requirements.txt
fastapi
strawberry-graphql[fastapi]
oracledb
python-dotenv
uvicorn
pytest
pytest-asyncio
httpx
